cilium: encryption route table need to account for tunnel headers
[ upstream commit 25a890ca67968f06ab5667159a7ea4dbbbce4d04 ]

The encryption table is used after encrypting packet. Here the packet is
encrypted but in the tunnel case must still account for encap overhead.
To ensure this set the mtu of routes here subtracted by the tunnel
overhead.

After this patch on a typical deployment with MTU=1500.

$ kubectl exec -ti -n kube-system cilium-9k9bk -- ip r s t 200
local 10.26.0.0/16 dev cilium_vxlan proto 50 scope host
10.250.0.0/16 dev cilium_host mtu 1450

Signed-off-by: John Fastabend <john.fastabend@gmail.com>
Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>