k8s: Disable several CiliumEndpoint status sections by default
The PATCH'ing of the status field of the CiliumEndpoint is often the bottleneck
in scalability. However, most of the frequently changing status sections
provided are not required for operation and only provide additional visibility.

Disable several CiliumEndpoint status sections by default to improve the
scalability of the CRD-backed identity allocation mode. Introduce a flag
--endpoint-status to re-enable individual status sections.

Signed-off-by: Thomas Graf <thomas@cilium.io>